Digit Recognizer/requirements.txt
joblib==1.0.1
matplotlib==3.4.3
numpy==1.22.4
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.1.1
streamlit==1.10.0
streamlit_drawable_canvas==0.9.1
